- Added more debug output to data abort handler - Added bit more debug output to event_init - Fixed possible implicit casting in strtoul - Fixed alignment calculation overflow within early heap - Used __builtin_assume_aligned to suppress wrong error message by compiler in heap - Added more debug output to early heap allocation - Added own library set flag macro for autotools - Added bytewise memset and memcpy to iomem since memcpy and memset may fail on device memory - Fixed EARLY_STARTUP_PRINT errors within iomem rpc init - Added check for copy_size to be smaller than PAGE_SIZE when performing mailbox call

+++ bolthur/kernel/lib/libc/stdlib/strtoul.c
@@ -60,12 +60,12 @@ unsigned long int strtoul(
       if ( *str - 'A' + 10 > base - 1 ) {
         break;
       }
-      digit = *str - 'A' + 10;
+      digit = (unsigned long int)(*str - 'A' + 10);
     } else if ( islower( *str ) ) {
       if ( *str - 'a' + 10 > base - 1 ) {
         break;
       }
-      digit = *str - 'a' + 10;
+      digit = (unsigned long int)(*str - 'a' + 10);
     } else {
       break;
     }
